Refactor tests; Extract mock objects; Create abstract test class
Preparatory task to make tests and test classes re-usable for new
components.
